[minor] Integer type consistency and fixing some warnings.

+++ Vulkan_Engine/src/Engine/RenderLoopStatistics.h
@@ -3,10 +3,10 @@
 
 struct FrameStats
 {
-	uint32_t pipelineCount;
-	uint32_t descriptorSetCount;
-	uint32_t drawCallCount;
+	size_t pipelineCount;
+	size_t descriptorSetCount;
+	size_t drawCallCount;
 
-	uint32_t frameNumber;
+	size_t frameNumber;
 	int64_t renderLoop_ms;
 };
